# resource/API/AZAZ音视频.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AZAZ音视频",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/azaz/interact/follow/list": {
      "get": {
        "summary": "关注列表查询",
        "x-apifox-folder": "社交模块",
        "x-apifox-status": "testing",
        "deprecated": false,
        "description": "查询用户关注列表，只能查询自己的关注列表",
        "tags": [
          "社交模块"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "",
            "required": true,
            "example": "rbh2559gtd7wrecknhxqmiedlwumjuhq",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "title": "返回码",
                      "description": "2成功，其他错误"
                    },
                    "message": {
                      "type": "string",
                      "title": "错误信息",
                      "description": "返回错误的情况向用户展示该错误信息",
                      "nullable": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/userInfo"
                      },
                      "title": "通用返回对象",
                      "description": "可以为空",
                      "nullable": true
                    }
                  },
                  "x-apifox-orders": [
                    "code",
                    "message",
                    "data"
                  ],
                  "required": [
                    "code",
                    "message",
                    "data"
                  ],
                  "x-apifox-refs": {},
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3472416/apis/api-120114872-run"
      }
    },
...
    "/azaz/interact/follow/friends": {
      "get": {
        "summary": "查询所有互关好友",
        "x-apifox-folder": "社交模块",
        "x-apifox-status": "testing",
        "deprecated": false,
        "description": "给互关用户分享视频，用户分享时选择要分享的互关用户，此时调用此接口",
        "tags": [
          "社交模块"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "",
            "required": true,
            "example": "rbh2559gtd7wrecknhxqmiedlwumjuhq",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "title": "返回码",
                      "description": "2操作成功，其他错误"
                    },
                    "message": {
                      "type": "string",
                      "title": "错误信息",
                      "description": "返回错误的情况向用户展示该错误信息",
                      "nullable": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/userInfo"
                      },
                      "title": "通用返回对象",
                      "description": "没有数据是为Null",
                      "nullable": true
                    }
                  },
                  "x-apifox-orders": [
                    "01HE2DYT79FFH4G5M6PX988YFR"
                  ],
                  "x-apifox-refs": {
                    "01HE2DYT79FFH4G5M6PX988YFR": {
                      "$ref": "#/components/schemas/Response"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ],
                  "x-apifox-ignore-properties": [
                    "code",
                    "message",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3472416/apis/api-120611918-run"
      }
    },
    "/azaz/interact/follow/ifFollow": {
      "get": {
        "summary": "查询是否关注",
        "x-apifox-folder": "社交模块",
        "x-apifox-status": "testing",
        "deprecated": false,
        "description": "查询第一个用户是否关注第二个",
        "tags": [
          "社交模块"
        ],
        "parameters": [
          {
            "name": "firstUser",
            "in": "query",
            "description": "第一个用户",
            "required": true,
            "example": "2",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "secondUser",
            "in": "query",
            "description": "第二个用户",
            "required": true,
            "example": "2",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "",
            "required": true,
            "example": "a7rs283hnj4qns8y8qksv3rqc8impokq",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "title": "返回码",
                      "description": "2操作成功，其他错误"
                    },
                    "message": {
                      "type": "string",
                      "title": "错误信息",
                      "description": "返回错误的情况向用户展示该错误信息",
                      "nullable": true
                    },
                    "data": {
                      "type": "boolean",
                      "title": "通用返回对象",
                      "description": "是否互关",
                      "nullable": true
                    }
                  },
                  "x-apifox-orders": [
                    "01HE2A1XQ9TCCD7YFFQ6Y1R3AP"
                  ],
                  "required": [
                    "code",
                    "message",
                    "data"
                  ],
                  "x-apifox-refs": {
                    "01HE2A1XQ9TCCD7YFFQ6Y1R3AP": {
                      "$ref": "#/components/schemas/Response"
                    }
                  },
                  "x-apifox-ignore-properties": [
                    "code",
                    "message",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3472416/apis/api-120611963-run"
      }
    },
...
    "/azaz/interact/follow/do": {
      "post": {
        "summary": "关注用户与取消关注",
        "x-apifox-folder": "社交模块",
        "x-apifox-status": "testing",
        "deprecated": false,
        "description": "在用户首页或者关注列表中可以选择关注用户或者取关",
        "tags": [
          "社交模块"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "操作对象",
            "required": true,
            "example": "1",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "操作类型 1-关注 0-取消关注",
            "required": true,
            "example": "1",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "请求头",
            "required": true,
            "example": "a4v05m4ato5yxyisuzyd31ebht88m0zv",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "title": "返回码",
                      "description": "2操作成功，其他错误"
                    },
                    "message": {
                      "type": "string",
                      "title": "错误信息",
                      "description": "返回错误的情况向用户展示该错误信息",
                      "nullable": true
                    },
                    "data": {
                      "type": "null",
                      "title": "通用返回对象",
                      "description": "没有数据"
                    }
                  },
                  "x-apifox-orders": [
                    "01HE2M16YR3B4FE418G532S6R7"
                  ],
                  "x-apifox-refs": {
                    "01HE2M16YR3B4FE418G532S6R7": {
                      "$ref": "#/components/schemas/Response"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ],
                  "x-apifox-ignore-properties": [
                    "code",
                    "message",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3472416/apis/api-121135874-run"
      }
    },
...
    "/azaz/interact/follow/num": {
      "get": {
        "summary": "查询关注数",
        "x-apifox-folder": "社交模块",
        "x-apifox-status": "testing",
        "deprecated": false,
        "description": "",
        "tags": [
          "社交模块"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "用户id",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "登录凭证",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "title": "返回码",
                      "description": "2操作成功，其他错误"
                    },
                    "message": {
                      "type": "string",
                      "title": "错误信息",
                      "description": "返回错误的情况向用户展示该错误信息",
                      "nullable": true
                    },
                    "data": {
                      "type": "integer",
                      "title": "通用返回对象",
                      "description": "关注人数",
                      "nullable": true
                    }
                  },
                  "x-apifox-orders": [
                    "01HE5CMT24KP1E2NTN8G1AGSR7"
                  ],
                  "required": [
                    "code",
                    "message",
                    "data"
                  ],
                  "x-apifox-refs": {
                    "01HE5CMT24KP1E2NTN8G1AGSR7": {
                      "$ref": "#/components/schemas/Response"
                    }
                  },
                  "x-apifox-ignore-properties": [
                    "code",
                    "message",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3472416/apis/api-121487116-run"
      }
    },
    "/azaz/interact/follow/fansNum": {
      "get": {
        "summary": "查询粉丝数",
        "x-apifox-folder": "社交模块",
        "x-apifox-status": "testing",
        "deprecated": false,
        "description": "",
        "tags": [
          "社交模块"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "用户id",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "登录凭证",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "title": "返回码",
                      "description": "2操作成功，其他错误"
                    },
                    "message": {
                      "type": "string",
                      "title": "错误信息",
                      "description": "返回错误的情况向用户展示该错误信息",
                      "nullable": true
                    },
                    "data": {
                      "type": "integer",
                      "title": "通用返回对象",
                      "description": "粉丝数",
                      "nullable": true
                    }
                  },
                  "x-apifox-orders": [
                    "01HE5CNFV8AFH2JSQCQ950JC6X"
                  ],
                  "x-apifox-refs": {
                    "01HE5CNFV8AFH2JSQCQ950JC6X": {
                      "$ref": "#/components/schemas/Response"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ],
                  "x-apifox-ignore-properties": [
                    "code",
                    "message",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3472416/apis/api-121487161-run"
      }
    },
...
    "/azaz/interact/follow/ifFollowEachOther": {
      "get": {
        "summary": "查询是否互相关注",
        "x-apifox-folder": "社交模块",
        "x-apifox-status": "testing",
        "deprecated": false,
        "description": "查询两个用户是否互相关注",
        "tags": [
          "社交模块"
        ],
        "parameters": [
          {
            "name": "firstUser",
            "in": "query",
            "description": "第一个用户的id",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "secondUser",
            "in": "query",
            "description": "第二个用户的id",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "title": "返回码",
                      "description": "2操作成功，其他错误"
                    },
                    "message": {
                      "type": "string",
                      "title": "错误信息",
                      "description": "返回错误的情况向用户展示该错误信息",
                      "nullable": true
                    },
                    "data": {
                      "type": "boolean",
                      "title": "通用返回对象",
                      "description": "两个用户是否互相关注",
                      "nullable": true
                    }
                  },
                  "x-apifox-orders": [
                    "01HED9SQCD5WTNT0H8TT5BYJH5"
                  ],
                  "required": [
                    "code",
                    "message",
                    "data"
                  ],
                  "x-apifox-refs": {
                    "01HED9SQCD5WTNT0H8TT5BYJH5": {
                      "$ref": "#/components/schemas/Response"
                    }
                  },
                  "x-apifox-ignore-properties": [
                    "code",
                    "message",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3472416/apis/api-122247336-run"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "userInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "用户id"
          },
          "username": {
            "type": "string",
            "description": "用户昵称"
          },
          "image": {
            "type": "string",
            "description": "用户头像"
          },
          "signature": {
            "type": "string",
            "description": "个性签名"
          }
        },
        "title": "用户基本信息",
        "x-apifox-orders": [
          "id",
          "username",
          "image",
          "signature"
        ],
        "required": [
          "username",
          "image",
          "id",
          "signature"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": ""
      },
...
      "Response": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "title": "返回码",
            "description": "2操作成功，其他错误"
          },
          "message": {
            "type": "string",
            "title": "错误信息",
            "description": "返回错误的情况向用户展示该错误信息",
            "nullable": true
          },
          "data": {
            "type": "object",
            "properties": {},
            "title": "通用返回对象",
            "description": "没有数据是为Null",
            "x-apifox-orders": [],
            "x-apifox-ignore-properties": [],
            "nullable": true
          }
        },
        "title": "通用返回对象",
        "x-apifox-orders": [
          "code",
          "message",
          "data"
        ],
        "required": [
          "code",
          "message",
          "data"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": ""
      },
...
    }
  },
  "servers": []
}